Update code docs and refactor
