+++ Tests/Shared.Specs/StringAssertionSpecs.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
+using FluentAssertions.Primitives;
 using Xunit;
 using Xunit.Sdk;
-using static FluentAssertions.Primitives.StringAssertions;
 
 namespace FluentAssertions.Specs
 {
